refactor: split the bot integration page into racks and dialogs
